+++ src/main/kotlin/kr/jiasoft/hiteen/feature/play/dto/RankingView.kt
@@ -2,6 +2,7 @@ package kr.jiasoft.hiteen.feature.play.dto
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import io.swagger.v3.oas.annotations.media.Schema
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -39,7 +40,7 @@ data class RankingView(
     val assetUid: String?,
 
     @field:Schema(description = "점수", example = "63.12 -> 01:03:12")
-    val score: Double,
+    val score: BigDecimal,
 
     @field:Schema(description = "시도 횟수", example = "2")
     val tryCount: Long? = null,
